#1 Sim speed fix
Moved the majority of process to a background thread.
Testing show that this appeared to be both stable, and effective at
reducing the latency issue.
My own testing showed that in a busy game world, turning on 1 solar
sensor dropped the sim speed from 0.7 to 0.2, and made it practically
impossible to play.
I have also optimized the crap out of the sun position calculation. It
cannot be improved any more. The Server and Client can still get out of
sync, but this is the Session.Elapsed Time, and it is out of my hands.
If Keen ever make it possible to change the Time Of Day, this will need
re-writing.